Refactor to collapse weird file structure
