+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "Marlea"
+name = "marlea"
 version = "0.1.0"
 edition = "2021"
