--- src/lib/marlea_engine/trial/reaction_network/tests.rs
@@ -1,9 +1,0 @@
-
-#[cfg(test)]
-mod tests {
-
-    use std::collections::{HashSet, HashMap};
-    use crate::marlea_engine::*;
-
-}
-
