+++ src/supported_file_type.rs
@@ -1,5 +1,5 @@
 
-use crate::marlea_engine::{reaction_network::reaction::{Reaction, term::{Term, species::Species}}};
+use crate::reaction_network::reaction::{Reaction, term::{Term, species::Species}};
 use std::{fs::File};
 use csv::ReaderBuilder; 
 use std::io::Write;
